Improved Cross-Platform Structure
Separated out the detection for different platforms.
