+++ rust-proxy/src/labview/install_detection_linux.rs
@@ -1,0 +1,8 @@
+
+
+use super::installs::{SystemLabviewInstalls, LabviewInstallError};
+
+/// Scan the system for LabVIEW installs and return their details.
+pub fn detect_installations() -> Result<SystemLabviewInstalls, LabviewInstallError> {
+    Ok(SystemLabviewInstalls::new())
+}
